Refactor LuckPerms integration and update mute compatibility providers for improved structure and support
