+++ src/main/java/off/szymon/vMessage/compatibility/mute/MutePluginCompatibilityProvider.java
@@ -1,4 +1,4 @@
-package off.szymon.vMessage.mute;
+package off.szymon.vMessage.compatibility.mute;
 
 import com.velocitypowered.api.proxy.Player;
